(cli-plugin) Use nodemon for server auto reload
For #10
