+++ packages/@uvue/vue-cli-plugin-ssr/package.json
@@ -52,6 +52,7 @@
     "lodash": "^4.17.11",
     "micromatch": "^3.1.10",
     "node-mocks-http": "^1.7.3",
+    "nodemon": "^1.18.7",
     "prettier": "^1.14.3",
     "querystring": "^0.2.0",
     "recursive-readdir": "^2.2.2",
